+++ src/core/request/engines/fetch.ts
@@ -12,12 +12,13 @@ import AbortController from 'abort-controller';
 //#endif
 
 import AbortablePromise from 'core/promise/abortable';
+import { isOnline } from 'core/net';
+
 import Response, { ResponseTypeValue } from 'core/request/response';
 import RequestError from 'core/request/error';
 
 import { convertDataToSend } from 'core/request/engines/helpers';
 import type { RequestEngine, NormalizedCreateRequestOptions } from 'core/request/interface';
-import { isOnline } from 'core/net';
 
 /**
  * Creates request by using the fetch API with the specified parameters and returns a promise