+++ src/core/request/engines/node.ts
@@ -9,11 +9,13 @@
 import got, { Options, CancelableRequest, Response as GotResponse } from 'got';
 
 import AbortablePromise from 'core/promise/abortable';
+import { isOnline } from 'core/net';
+
 import Response from 'core/request/response';
 import RequestError from 'core/request/error';
 
 import { convertDataToSend } from 'core/request/engines/helpers';
-import type { RequestEngine } from 'core/request/interface';
+import type { RequestEngine, NormalizedCreateRequestOptions } from 'core/request/interface';
 
 /**
  * Creates request by using node.js with the specified parameters and returns a promise
@@ -69,7 +71,16 @@ const request: RequestEngine = (params) => {
 		normalizedOpts.responseType = 'buffer';
 	}
 
-	return new AbortablePromise<Response>((resolve, reject, onAbort) => {
+	return new AbortablePromise<Response>(async (resolve, reject, onAbort) => {
+		const
+			{status} = await AbortablePromise.resolve(isOnline(), p.parent);
+
+		if (!status) {
+			return reject(new RequestError('offline', {
+				request: <NormalizedCreateRequestOptions>normalizedOpts
+			}));
+		}
+
 		const
 			request = <CancelableRequest<GotResponse>>got(p.url, normalizedOpts);